# TS: cache type IDs on type objects