@asger-semmle 2018-11-30
After re-running vscode: 

|   | before | after | difference |
--|-|-|-|
| vscode | 231s | 223s | 9s (3%) |