@asger-semmle 2018-12-03
Turns out the tests are right, so I'll just write up the story here for posterity.

Types can be extracted as "shallow" or "full". A fully-extracted type also has the type of its properties extracted. A type that is directly referenced from the AST gets fully extracted, whereas a type is shallow it is only reachable through the unfolding of another type's properties. A type can transition from shallow to full if it was previously seen in shallow context, but then later it is found to be referenced in full context (e.g. it is the type of an expression). We detect the need for this transition in the `getId` method.

However, when this transition was about to happen, the ID was simply found in the new cache and `getId` returned without doing the transition. To fix this, we would have to unfold the properties of the type again, to promote all the types that would now be reachable in full context. Previously, these checks just happened by calling `getTypeString`, due to its the recursive calls to `getId`.

Fixing this is beyond the out of scope of what I thought was just a low-hanging fruit, so for now I'll just close this PR.